requirements.txt
aniso8601==8.0.0
arrow==0.15.2
attrs==19.2.0
boto==2.49.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-cors
flask-restplus==0.13.0
Flask-Script==2.0.6
grpcio==1.24.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
matchbox-orm==0.2.6
mock==3.0.5
more-itertools==7.2.0
msgpack==0.6.2
pandas==0.25.3
passlib
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyfcm==1.4.7
pylint==2.4.4
pyrsistent==0.15.4
pytest==5.3.1
pytest-cov==2.8.1
pytest-mock==1.12.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
waitress==1.3.1
Werkzeug==0.16.0
zipp==0.6.0


